Improved resetSearchResults: do nothing when no rows found

+++ app/code/community/JeroenVermeulen/Solarium/Model/Resource/CatalogSearch/Fulltext.php
@@ -124,45 +124,48 @@ class JeroenVermeulen_Solarium_Model_Resource_CatalogSearch_Fulltext extends Mag
         $queryTable  = $this->getTable('catalogsearch/search_query');
         $querySelect = $read->select()->from( $queryTable, 'COUNT(*)' )->where( 'is_processed' );
         $queryCount  = intval( $read->fetchOne( $querySelect ) );
-        Mage::log( sprintf('Solarium resetSearchResults: clearing %d search queries', $queryCount), Zend_Log::DEBUG );
-        $queryPages  = ceil( $queryCount / $pageSize );
-        $querySql    = sprintf( 'UPDATE %s SET %s=0 WHERE %s=1 LIMIT %d',
-                                $read->quoteIdentifier( $queryTable ),
-                                $read->quoteIdentifier( 'is_processed' ),
-                                $read->quoteIdentifier( 'is_processed' ),
-                                $pageSize );
-        for ( $page=0; $page < $queryPages; $page++ ) {
-            try {
-                // It would be better to do this using a Varien or Zend object, but they don't support LIMIT on update.
-                $write->query( $querySql );
-            } catch ( Exception $e ) {
-                // This happens on busy sites because of deadlock
-                Mage::log( sprintf('Solarium resetSearchResults: Update error during search reindex, but we can continue: %s', $e->getMessage()),
-                           Zend_Log::ERR );
-                sleep(1);
-                $page--;
+        if ( $queryCount ) {
+            Mage::log( sprintf('Solarium resetSearchResults: clearing %d search queries', $queryCount), Zend_Log::DEBUG );
+            $queryPages  = ceil( $queryCount / $pageSize );
+            $querySql    = sprintf( 'UPDATE %s SET %s=0 WHERE %s=1 LIMIT %d',
+                                    $read->quoteIdentifier( $queryTable ),
+                                    $read->quoteIdentifier( 'is_processed' ),
+                                    $read->quoteIdentifier( 'is_processed' ),
+                                    $pageSize );
+            for ( $page=0; $page < $queryPages; $page++ ) {
+                try {
+                    // It would be better to do this using a Varien or Zend object, but they don't support LIMIT on update.
+                    $write->query( $querySql );
+                } catch ( Exception $e ) {
+                    // This happens on busy sites because of deadlock
+                    Mage::log( sprintf('Solarium resetSearchResults: Update error during search reindex, but we can continue: %s', $e->getMessage()),
+                               Zend_Log::ERR );
+                    sleep(1);
+                    $page--;
+                }
             }
         }
-
         $resultTable  = $this->getTable('catalogsearch/result');
         $resultSelect = $read->select()->from( $resultTable, 'COUNT(*)' );
         $resultCount  = intval( $read->fetchOne( $resultSelect ) );
-        Mage::log( sprintf('Solarium resetSearchResults: clearing %d search results', $queryCount), Zend_Log::DEBUG );
-        $resultPages  = ceil( $resultCount / $pageSize );
-        $resultSql    = sprintf( 'DELETE FROM %s LIMIT %d',
-                                 $read->quoteIdentifier( $resultTable ),
-                                 $pageSize );
-        for ( $page=0; $page < $resultPages; $page++ ) {
-            try {
-                // It would be better to do this using a Varien or Zend object, but they don't support LIMIT on delete.
-                $write->query( $resultSql );
-                break;
-            } catch ( Exception $e ) {
-                // This happens on busy sites because of deadlock
-                Mage::log( sprintf('Solarium resetSearchResults: Delete error during search reindex, but we can continue: %s', $e->getMessage()),
-                           Zend_Log::ERR );
-                sleep(1);
-                $page--;
+        if ( $resultCount ) {
+            Mage::log( sprintf('Solarium resetSearchResults: clearing %d search results', $queryCount), Zend_Log::DEBUG );
+            $resultPages  = ceil( $resultCount / $pageSize );
+            $resultSql    = sprintf( 'DELETE FROM %s LIMIT %d',
+                                     $read->quoteIdentifier( $resultTable ),
+                                     $pageSize );
+            for ( $page=0; $page < $resultPages; $page++ ) {
+                try {
+                    // It would be better to do this using a Varien or Zend object, but they don't support LIMIT on delete.
+                    $write->query( $resultSql );
+                    break;
+                } catch ( Exception $e ) {
+                    // This happens on busy sites because of deadlock
+                    Mage::log( sprintf('Solarium resetSearchResults: Delete error during search reindex, but we can continue: %s', $e->getMessage()),
+                               Zend_Log::ERR );
+                    sleep(1);
+                    $page--;
+                }
             }
         }
